requirements.txt
Django==5.1
rest_framework_simplejwt==0.0.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-filter==24.3
pyotp==2.9.0